Move more expression code to datafusion-expr
